moved AutoFlow to Parameterized class
