+++ GPflow/model.py
@@ -1,11 +1,10 @@
 from __future__ import print_function
-from .param import Parameterized
+from .param import Parameterized, AutoFlow
 from scipy.optimize import minimize, OptimizeResult
 import numpy as np
 import tensorflow as tf
 from . import hmc
 import sys
-from functools import wraps
 
 
 class ObjectiveWrapper(object):
@@ -30,78 +29,6 @@ class ObjectiveWrapper(object):
         else:
             print("Warning: inf or nan in gradient: replacing with zeros")
             return f, np.where(g_is_fin, g, 0.)
-
-
-class AutoFlow:
-    """
-    This decorator-class is designed for use on methods of the Model class
-    (below).
-
-    The idea is that methods that compute relevant quantities of the model
-    (such as predictions) can define a tf graph which we automatically run when
-    the (decorated) function is called.
-
-    The syntax looks like:
-
-    >>> class MyModel(Model):
-    >>>
-    >>>   @AutoFlow((tf.float64), (tf.float64))
-    >>>   def my_method(self, x, y):
-    >>>       #compute something, returning a tf graph.
-    >>>       return tf.foo(self.baz, x, y)
-
-    >>> m = MyModel()
-    >>> x = np.random.randn(3,1)
-    >>> y = np.random.randn(3,1)
-    >>> result = my_method(x, y)
-
-    Now the output of the method call is the _result_ of the computation,
-    equivalent to
-
-    >>> m = MyModel()
-    >>> x = np.random.randn(3,1)
-    >>> y = np.random.randn(3,1)
-    >>> x_tf = tf.placeholder(tf.float64)
-    >>> y_tf = tf.placeholder(tf.float64)
-    >>> with m.tf_mode():
-    >>>     graph = tf.foo(m.baz, x_tf, y_tf)
-    >>> result = m._session.run(graph,
-                                feed_dict={x_tf:x,
-                                           y_tf:y,
-                                           m._free_vars:m.get_free_state()})
-
-    Not only is the syntax cleaner, but multiple calls to the method will
-    result in the graph being constructed only once.
-
-    """
-
-    def __init__(self, *tf_arg_tuples):
-        # NB. TF arg_tuples is a list of tuples, each of which can be used to
-        # construct a tf placeholder.
-        self.tf_arg_tuples = tf_arg_tuples
-
-    def __call__(self, tf_method):
-        @wraps(tf_method)
-        def runnable(instance, *np_args):
-            storage_name = '_' + tf_method.__name__ + '_AF_storage'
-            if hasattr(instance, storage_name):
-                # the method has been compiled already, get things out of storage
-                storage = getattr(instance, storage_name)
-            else:
-                # the method needs to be compiled.
-                storage = {}  # an empty dict to keep things in
-                setattr(instance, storage_name, storage)
-                storage['free_vars'] = tf.placeholder(tf.float64, [None])
-                instance.make_tf_array(storage['free_vars'])
-                storage['tf_args'] = [tf.placeholder(*a) for a in self.tf_arg_tuples]
-                with instance.tf_mode():
-                    storage['tf_result'] = tf_method(instance, *storage['tf_args'])
-                storage['session'] = tf.Session()
-            feed_dict = dict(zip(storage['tf_args'], np_args))
-            feed_dict[storage['free_vars']] = instance.get_free_state()
-            return storage['session'].run(storage['tf_result'], feed_dict=feed_dict)
-
-        return runnable
 
 
 class Model(Parameterized):
@@ -175,7 +102,7 @@ class Model(Parameterized):
             opt_step = None
         else:
             opt_step = optimizer.minimize(self._minusF,
-                                              var_list=[self._free_vars])
+                                          var_list=[self._free_vars])
         init = tf.initialize_all_variables()
         self._session.run(init)
 
@@ -193,14 +120,6 @@ class Model(Parameterized):
         self._needs_recompile = False
 
         return opt_step
-
-    def __setattr__(self, key, value):
-        Parameterized.__setattr__(self, key, value)
-        # delete any AutoFlow related graphs
-        if key == '_needs_recompile' and value:
-            for key in dir(self):
-                if key[0] == '_' and key[-11:] == '_AF_storage':
-                    delattr(self, key)
 
     @AutoFlow()
     def compute_log_prior(self):
